requirements.txt
fastapi[all]
pandas
pydantic
python-dotenv
httpx
pydantic-settings
#qdrant-client
requests
#sentence-transformers
sqlalchemy